+++ cucumber-react/javascript/src/components/gherkin/Rule.tsx
@@ -4,7 +4,6 @@ import Description from './Description'
 import Scenario from './Scenario'
 import { messages } from 'cucumber-messages'
 import Background from './Background'
-import { H2, Indent, StepText, Section } from './html'
 import IRule = messages.GherkinDocument.Feature.FeatureChild.IRule
 
 interface IProps {
@@ -13,11 +12,12 @@ interface IProps {
 
 const Rule: React.FunctionComponent<IProps> = ({ rule }) => {
   return (
-    <Section>
-      <H2>
-        <Keyword>{rule.keyword}:</Keyword> <StepText>{rule.name}</StepText>
-      </H2>
-      <Indent>
+    <section>
+      <h2>
+        <Keyword>{rule.keyword}:</Keyword>{' '}
+        <span className="step-text">{rule.name}</span>
+      </h2>
+      <div className="indent">
         <Description description={rule.description} />
         {(rule.children || []).map((child, index) => {
           if (child.background) {
@@ -28,8 +28,8 @@ const Rule: React.FunctionComponent<IProps> = ({ rule }) => {
             throw new Error('Expected background or scenario')
           }
         })}
-      </Indent>
-    </Section>
+      </div>
+    </section>
   )
 }