+++ cucumber-react/javascript/src/components/gherkin/Step.tsx
@@ -3,34 +3,9 @@ import DataTable from './DataTable'
 import Keyword from './Keyword'
 import DocString from './DocString'
 import { messages } from 'cucumber-messages'
-import styled from 'styled-components'
 import statusColor from './statusColor'
-import { StepParam, H3, Indent, StepText, IStatusProps } from './html'
 import CucumberQueryContext from '../../CucumberQueryContext'
 import UriContext from '../../UriContext'
-
-const StepLi = styled.li`
-  padding: 0.5em;
-  margin-left: 0;
-  margin-top: 0;
-  border-bottom: 1px #ccc solid;
-  border-left: 1px #ccc solid;
-  border-right: 1px #ccc solid;
-  background-color: ${(props: IStatusProps) => statusColor(props.status).hex()};
-
-  &:nth-child(1) {
-    border-top: 1px #ccc solid;
-  }
-`
-
-const ErrorMessage = styled.pre`
-  padding: 0.5em;
-  background-color: ${(props: IStatusProps) =>
-    statusColor(props.status)
-      .darken(0.1)
-      .hex()};
-  overflow: scroll;
-`
 
 interface IProps {
   step: messages.GherkinDocument.Feature.IStep
@@ -67,57 +42,85 @@ const Step: React.FunctionComponent<IProps> = ({
         plain = step.text.slice(offset, argument.group.start)
         if (plain.length > 0) {
           stepTextElements.push(
-            <StepText key={`plain-${index}`}>{plain}</StepText>
+            <span className="step-text" key={`plain-${index}`}>
+              {plain}
+            </span>
           )
         }
         const arg = argument.group.value
         if (arg.length > 0) {
           stepTextElements.push(
-            <StepParam key={`bold-${index}`} status={status}>
+            <span
+              className="step-param"
+              key={`bold-${index}`}
+              style={{
+                backgroundColor: statusColor(status)
+                  .darken(0.1)
+                  .hex(),
+              }}
+            >
               {arg}
-            </StepParam>
+            </span>
           )
         }
         offset += plain.length + arg.length
       })
       plain = step.text.slice(offset)
       if (plain.length > 0) {
-        stepTextElements.push(<StepText key={`plain-rest`}>{plain}</StepText>)
+        stepTextElements.push(
+          <span className="step-text" key={`plain-rest`}>
+            {plain}
+          </span>
+        )
       }
     } else if (stepMatchArgumentsLists.length === 2) {
       // Step is ambiguous
       stepTextElements.push(
-        <StepText key={`plain-ambiguous`}>{step.text}</StepText>
+        <span className="step-text" key={`plain-ambiguous`}>
+          {step.text}
+        </span>
       )
     } else {
       // Step is undefined
       stepTextElements.push(
-        <StepText key={`plain-undefined`}>{step.text}</StepText>
+        <span className="step-text" key={`plain-undefined`}>
+          {step.text}
+        </span>
       )
     }
   } else {
     // Step is from scenario with examples, and has <> placeholders.
     stepTextElements.push(
-      <StepText key={`plain-placeholders`}>{step.text}</StepText>
+      <span className="step-text" key={`plain-placeholders`}>
+        {step.text}
+      </span>
     )
   }
 
   return (
-    <StepLi status={status}>
-      <H3>
+    <li className="step" style={{ backgroundColor: statusColor(status).hex() }}>
+      <h3>
         <Keyword>{step.keyword}</Keyword>
         {stepTextElements}
-      </H3>
-      <Indent>
+      </h3>
+      <div className="indent">
         {step.dataTable && <DataTable dataTable={step.dataTable} />}
         {step.docString && <DocString docString={step.docString} />}
-      </Indent>
+      </div>
       {resultsWithMessage.map((result, i) => (
-        <ErrorMessage key={i} status={status}>
+        <pre
+          className="error-message"
+          key={i}
+          style={{
+            backgroundColor: statusColor(status)
+              .darken(0.1)
+              .hex(),
+          }}
+        >
           {result.message}
-        </ErrorMessage>
+        </pre>
       ))}
-    </StepLi>
+    </li>
   )
 }