+++ cucumber-react/javascript/src/components/gherkin/TableBody.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { messages } from 'cucumber-messages'
-import { Td } from './html'
+import isNumber from './isNumber'
 
 interface IProps {
   rows: messages.GherkinDocument.Feature.ITableRow[]
@@ -12,7 +12,12 @@ const TableBody: React.FunctionComponent<IProps> = ({ rows }) => {
       {rows.map((row, i) => (
         <tr key={i}>
           {(row.cells || []).map((cell, j) => (
-            <Td key={j}>{cell.value}</Td>
+            <td
+              key={j}
+              style={{ textAlign: isNumber(cell.value) ? 'right' : 'left' }}
+            >
+              {cell.value}
+            </td>
           ))}
         </tr>
       ))}