+++ cucumber-react/javascript/src/components/gherkin/Tags.tsx
@@ -1,28 +1,23 @@
 import React from 'react'
 import { messages } from 'cucumber-messages'
-import Tag from './Tag'
 import ITag = messages.GherkinDocument.Feature.ITag
-import styled from 'styled-components'
 
 interface IProps {
   tags: ITag[]
 }
-
-const TagList = styled.ul`
-  padding: 0;
-  margin-bottom: 0;
-`
 
 const Tags: React.FunctionComponent<IProps> = ({ tags }) => {
   if (!tags) {
     return null
   }
   return (
-    <TagList>
+    <ul className="tag-list">
       {tags.map((tag, index) => (
-        <Tag key={index}>{tag.name}</Tag>
+        <li className="tag" key={index}>
+          {tag.name}
+        </li>
       ))}
-    </TagList>
+    </ul>
   )
 }