+++ cucumber-react/javascript/stories/index.stories.tsx
@@ -5,6 +5,8 @@ import StepList from '../src/components/gherkin/StepList'
 import GherkinDocumentList from '../src/components/app/GherkinDocumentList'
 import CucumberQuery from 'cucumber-query'
 import CucumberQueryContext from '../src/CucumberQueryContext';
+import '../src/styles/react-accessible-accordion.css'
+import '../src/styles/styles.css'
 
 // @ts-ignore
 import ndjson from '../testdata/all.ndjson'